LPS-198452 Inline class OrganizationTypeConfigurationWrapper in class OrganizationTypesSettingsImpl
